+++ packages/components/src/accessible-popover/README.md
@@ -1,15 +1,15 @@
-# Popover
+# AccessiblePopover
 
 <div class="callout callout-alert">
 This feature is still experimental. “Experimental” means this is an early implementation subject to drastic and breaking changes.
 </div>
 
-`Popover` is a component to render a floating content modal. It is similar in purpose to a tooltip, but renders content of any sort, not only simple text.
+`AccessiblePopover` is a component to render a floating content modal. It is similar in purpose to a tooltip, but renders content of any sort, not only simple text.
 
 ## Usage
 
 ```jsx
-import { Button, Popover, View, Text } from '@wordpress/components/ui';
+import { Button, __experimentalAccessiblePopover as Popover, View, Text } from '@wordpress/components';
 
 function Example() {
 	return (
@@ -21,3 +21,7 @@ function Example() {
 	);
 }
 ```
+
+## Props
+
+TODO
