+++ packages/components/src/accessible-popover/content.js
@@ -11,16 +11,16 @@ import { Popover as ReakitPopover } from 'reakit';
 /**
  * Internal dependencies
  */
-import { Card } from '../../card';
-import { View } from '../../view';
+import { Card } from '../card';
+import { View } from '../view';
 import { usePopoverContext } from './context';
 import * as styles from './styles';
-import { contextConnect, useContextSystem } from '../context';
+import { contextConnect, useContextSystem } from '../ui/context';
 
 /**
  *
- * @param {import('../context').PolymorphicComponentProps<import('./types').ContentProps, 'div'>} props
- * @param {import('react').Ref<any>}                                                              forwardedRef
+ * @param {import('../ui/context').PolymorphicComponentProps<import('./types').ContentProps, 'div'>} props
+ * @param {import('react').Ref<any>}                                                                 forwardedRef
  */
 function PopoverContent( props, forwardedRef ) {
 	const {
