+++ packages/components/src/accessible-popover/index.js
@@ -1,0 +1,2 @@
+export { default as AccessiblePopover } from './component';
+export * from './context';
