+++ packages/components/src/accessible-popover/styles.js
@@ -9,9 +9,9 @@ import { css } from '@emotion/css';
 /**
  * Internal dependencies
  */
-import { CardBody } from '../../card';
-import * as ZIndex from '../../utils/z-index';
-import CONFIG from '../../utils/config-values';
+import { CardBody } from '../card';
+import * as ZIndex from '../utils/z-index';
+import CONFIG from '../utils/config-values';
 
 export const PopoverContent = css`
 	z-index: ${ ZIndex.Popover };
