--- packages/components/src/ui/popover/index.js
@@ -1,2 +1,0 @@
-export { default as Popover } from './component';
-export * from './context';
